centralize block type defs
