+++ src/plugins/spark/features/ActuatorLogic/types.ts
@@ -1,0 +1,6 @@
+
+export interface ExpressionError {
+  index: number;
+  message: string;
+  indicator: string;
+}
