--- src/plugins/spark/types/block-types-local.ts
@@ -1,20 +1,0 @@
-export interface ChannelMapping {
-  id: string;
-  nid: number;
-  name: string;
-}
-
-export interface DisplayOpts {
-  color: string;
-  name: string;
-  pos?: number;
-  unique: boolean;
-  showNotify: boolean;
-  showDialog: boolean;
-}
-
-export interface ExpressionError {
-  index: number;
-  message: string;
-  indicator: string;
-}
