+++ src/plugins/spark/types/index.ts
@@ -1,6 +1,2 @@
-export * from './block-types';
 export * from './spark-types';
-export * from './constraint-types';
-export * from './block-data-enums';
-export * from './block-types-local';
-export * from './block-data-types';
+export * from './shared';
