+++ src/plugins/spark/types/shared/index.ts
@@ -1,0 +1,2 @@
+export * from './block-enums';
+export * from './block-types';
